+++ src/main/java/io/jenkins/plugins/analysis/core/model/IssueStream.java
@@ -111,6 +111,7 @@ public class IssueStream {
      * Default {@link Converter} implementation for XStream that does interning scoped to one unmarshalling.
      */
     private static final class SeverityConverter implements Converter {
+        @SuppressWarnings("PMD.NullAssignment")
         @Override
         public void marshal(final Object source, final HierarchicalStreamWriter writer,
                 final MarshallingContext context) {