+++ src/main/java/io/jenkins/plugins/analysis/core/scm/NullBlamer.java
@@ -9,10 +9,12 @@ import edu.hm.hafner.analysis.Report;
  */
 public class NullBlamer implements Blamer {
     private static final long serialVersionUID = -338286497290046470L;
+    /** A message that explains that no blaming information will be collected. **/
+    static final String BLAMING_SKIPPED = "Skipping blaming as requested in the job configuration";
 
     @Override
     public Blames blame(final Report report) {
-        report.logInfo("Skipping blaming as requested in the job configuration");
+        report.logInfo(BLAMING_SKIPPED);
         return new Blames();
     }
 }