+++ src/main/java/io/jenkins/plugins/analysis/warnings/axivion/RemoteAxivionDashboard.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import org.apache.http.HttpResponse;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.CredentialsProvider;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.BasicCredentialsProvider;
@@ -22,40 +23,34 @@ import net.sf.json.JSONObject;
  * Represents an actual dashboard connection to retrieve violations via http.
  */
 class RemoteAxivionDashboard implements AxivionDashboard {
-
     private static final String X_AXIVION_USER_AGENT = "x-axivion-user-agent";
     private static final String API_USER_AGENT = "ApiClient/6.9.3";
     private static final String FALL_USER_AGENT = "AxivionEclipsePlugin/6.9.2";
+    private static final int HTTP_STATUS_OK = 200;
 
     private final String projectUrl;
     private final UsernamePasswordCredentials credentials;
 
-    RemoteAxivionDashboard(
-            final String projectUrl,
-            final UsernamePasswordCredentials credentials) {
+    RemoteAxivionDashboard(final String projectUrl, final UsernamePasswordCredentials credentials) {
         this.projectUrl = projectUrl;
         this.credentials = credentials;
     }
 
     @Override
     public JSONObject getIssues(final AxIssueKind kind) {
-        final org.apache.http.client.CredentialsProvider credentialsProvider =
-                new BasicCredentialsProvider();
-
+        CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
         credentialsProvider.setCredentials(AuthScope.ANY, credentials);
 
         try (CloseableHttpClient client =
-                     HttpClients.custom()
-                             .setDefaultCredentialsProvider(credentialsProvider)
-                             .build()) {
+                     HttpClients.custom().setDefaultCredentialsProvider(credentialsProvider).build()) {
 
-            HttpGet httpget = new HttpGet(this.projectUrl + "/issues?kind=" + kind);
+            HttpGet httpget = new HttpGet(projectUrl + "/issues?kind=" + kind);
             httpget.addHeader(new BasicHeader("Accept", "application/json"));
             BasicHeader userAgent = new BasicHeader(X_AXIVION_USER_AGENT, API_USER_AGENT);
             httpget.addHeader(userAgent);
 
             try (CloseableHttpResponse response = client.execute(httpget)) {
-                if (response.getStatusLine().getStatusCode() == 200) {
+                if (response.getStatusLine().getStatusCode() == HTTP_STATUS_OK) {
                     return convertToJson(response);
                 }
             }
@@ -68,15 +63,14 @@ class RemoteAxivionDashboard implements AxivionDashboard {
             }
         }
         catch (IOException e) {
-            throw new ParsingException("Cannot retrieve information from dashboard." + e.getMessage());
+            throw new ParsingException(e, "Cannot retrieve information from dashboard");
         }
         catch (JSONException e) {
-            throw new ParsingException("Invalid JSON response from dashboard." + e.getMessage());
+            throw new ParsingException(e, "Invalid JSON response from dashboard");
         }
     }
 
     private JSONObject convertToJson(final HttpResponse response) throws IOException {
-        final String result = EntityUtils.toString(response.getEntity());
-        return JSONObject.fromObject(result);
+        return JSONObject.fromObject(EntityUtils.toString(response.getEntity()));
     }
 }