+++ src/test/java/io/jenkins/plugins/analysis/core/charts/LineSeriesTest.java
@@ -3,10 +3,9 @@ package io.jenkins.plugins.analysis.core.charts;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
 
-import net.sf.json.JSONObject;
-
 import io.jenkins.plugins.analysis.core.charts.LineSeries.FilledMode;
 import io.jenkins.plugins.analysis.core.charts.LineSeries.StackedMode;
+import io.jenkins.plugins.analysis.core.util.JacksonFacade;
 
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.*;
 
@@ -23,14 +22,14 @@ class LineSeriesTest {
 
     @Test
     void shouldCreateLineSeriesStackedLines() {
-        JSONObject lineSeries = createLineSeries(StackedMode.STACKED, FilledMode.LINES);
+        String lineSeries = createLineSeries(StackedMode.STACKED, FilledMode.LINES);
 
-        assertThatJson(lineSeries).node("areaStyle").isEqualTo("{}");
+        assertThatJson(lineSeries).node("areaStyle").isNull();
         assertThatJson(lineSeries).node("stack").isEqualTo(STACKED);
         assertThatOtherPropertiesAreCorrectlySet(lineSeries);
     }
 
-    private void assertThatOtherPropertiesAreCorrectlySet(final JSONObject lineSeries) {
+    private void assertThatOtherPropertiesAreCorrectlySet(final String lineSeries) {
         assertThatJson(lineSeries).node("name").isEqualTo(SEVERITY);
         assertThatJson(lineSeries).node("type").isEqualTo(LINE);
         assertThatJson(lineSeries).node("data").isArray().hasSize(0);
@@ -38,16 +37,16 @@ class LineSeriesTest {
 
     @Test
     void shouldCreateLineSeriesLines() {
-        JSONObject lineSeries = createLineSeries(StackedMode.SEPARATE_LINES, FilledMode.LINES);
+        String lineSeries = createLineSeries(StackedMode.SEPARATE_LINES, FilledMode.LINES);
 
-        assertThatJson(lineSeries).node("areaStyle").isEqualTo("{}");
+        assertThatJson(lineSeries).node("areaStyle").isNull();
         assertThatJson(lineSeries).node("stack").isEqualTo(StringUtils.EMPTY);
         assertThatOtherPropertiesAreCorrectlySet(lineSeries);
     }
 
     @Test
     void shouldCreateLineSeriesStackedFilled() {
-        JSONObject lineSeries = createLineSeries(StackedMode.STACKED, FilledMode.FILLED);
+        String lineSeries = createLineSeries(StackedMode.STACKED, FilledMode.FILLED);
 
         assertThatJson(lineSeries).node("areaStyle").isEqualTo(new AreaStyle());
         assertThatJson(lineSeries).node("stack").isEqualTo(STACKED);
@@ -56,7 +55,7 @@ class LineSeriesTest {
 
     @Test
     void shouldCreateLineSeriesLinesFilled() {
-        JSONObject lineSeries = createLineSeries(StackedMode.SEPARATE_LINES, FilledMode.FILLED);
+        String lineSeries = createLineSeries(StackedMode.SEPARATE_LINES, FilledMode.FILLED);
 
         assertThatJson(lineSeries).node("areaStyle").isEqualTo(new AreaStyle());
         assertThatJson(lineSeries).node("stack").isEqualTo(StringUtils.EMPTY);
@@ -71,8 +70,7 @@ class LineSeriesTest {
         assertThatJson(lineSeries).node("data").isArray().hasSize(1).contains(22);
     }
 
-    private JSONObject createLineSeries(final StackedMode stacked, final FilledMode lines) {
-        LineSeries lineSeries = new LineSeries(SEVERITY, COLOR, stacked, lines);
-        return JSONObject.fromObject(lineSeries);
+    private String createLineSeries(final StackedMode stacked, final FilledMode lines) {
+        return new JacksonFacade().toJson(new LineSeries(SEVERITY, COLOR, stacked, lines));
     }
 }